+++ src/main/java/com/valuequo/buckswise/service/dto/MiscellenousDTO.java
@@ -2,22 +2,22 @@ package com.valuequo.buckswise.service.dto;
 
 public class MiscellenousDTO {
 	
-	private int userid;
+	private Long userid;
 	private String name;
 	private String value;
 	
 
-	public MiscellenousDTO(int userid, String name, String value) {
+	public MiscellenousDTO(Long userid, String name, String value) {
 		super();
 		this.userid = userid;
 		this.name = name;
 		this.value = value;
 	}
 
-	public int getUserid() {
+	public Long getUserid() {
 		return userid;
 	}
-	public void setUserid(int userid) {
+	public void setUserid(Long userid) {
 		this.userid = userid;
 	}
 	public String getName() {
@@ -32,7 +32,4 @@ public class MiscellenousDTO {
 	public void setValue(String value) {
 		this.value = value;
 	}
-	
-	
-
 }
